RAG/requirements.txt
flask
langchain-core
langchain-community
langchain-openai
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
